Temp

+++ src/main/java/org/work/secretsantabot/model/User.java
@@ -23,7 +23,7 @@ public class User {
     private String telegramUsername;
 
     @Column(nullable = false)
-    private int telegramChatId;
+    private long telegramChatId;
 
     private String authToken;
